+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.0](https://github.com/mittwald/flow/compare/1.0.16...1.1.0) (2026-08-31)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+## [1.0.16](https://github.com/mittwald/flow/compare/1.0.15...1.0.16) (2026-08-31)
+
+### Performance Improvements
+
+* **components:** stop shipping design-token build metadata to the browser ([#3007](https://github.com/mittwald/flow/issues/3007)) ([7d3afde](https://github.com/mittwald/flow/commit/7d3afde851c5f51f4a641e36dd549500e6fd759b))
+
 ## [1.0.15](https://github.com/mittwald/flow/compare/1.0.14...1.0.15) (2026-08-31)
 
 **Note:** Version bump only for package @mittwald/flow-project